steamguides.api: Add workaround for corrupted image reupload failure
Even after corrupted images are deleted, it is still not possible to
upload images with the same file name. Steam's backend simply returns
an empty array where the uploaded image data should be. Add an
automatic workaround for this circumstance by giving the uploaded
image a unique file name, which seems to make Steam's API happy.